Renaming package `event` to `events`
