+++ missilewars-plugin/src/main/java/de/butzlabben/missilewars/events/GameStartEvent.java
@@ -16,29 +16,18 @@
  * along with MissileWars.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package de.butzlabben.missilewars.event;
+package de.butzlabben.missilewars.events;
 
 import de.butzlabben.missilewars.game.Game;
-import lombok.Getter;
-import org.bukkit.entity.Player;
-import org.bukkit.event.Cancellable;
 import org.bukkit.event.HandlerList;
-import org.bukkit.event.player.PlayerEvent;
 import org.jetbrains.annotations.NotNull;
 
-/**
- * Get's called, when a player has already entered an arena
- */
-@Getter
-public class PlayerArenaJoinEvent extends PlayerEvent implements Cancellable {
+public class GameStartEvent extends GameEvent {
 
     public final static HandlerList handlers = new HandlerList();
-    private final Game game;
-    private boolean cancelled;
 
-    public PlayerArenaJoinEvent(Player who, Game game) {
-        super(who);
-        this.game = game;
+    public GameStartEvent(Game game) {
+        super(game);
     }
 
     public static HandlerList getHandlerList() {
@@ -49,15 +38,5 @@ public class PlayerArenaJoinEvent extends PlayerEvent implements Cancellable {
     @NotNull
     public HandlerList getHandlers() {
         return handlers;
-    }
-
-    @Override
-    public boolean isCancelled() {
-        return cancelled;
-    }
-
-    @Override
-    public void setCancelled(boolean b) {
-        cancelled = b;
     }
 }
