+++ missilewars-plugin/src/main/java/de/butzlabben/missilewars/events/PlayerArenaJoinEvent.java
@@ -16,17 +16,29 @@
  * along with MissileWars.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package de.butzlabben.missilewars.event;
+package de.butzlabben.missilewars.events;
 
 import de.butzlabben.missilewars.game.Game;
+import lombok.Getter;
+import org.bukkit.entity.Player;
+import org.bukkit.event.Cancellable;
 import org.bukkit.event.HandlerList;
+import org.bukkit.event.player.PlayerEvent;
+import org.jetbrains.annotations.NotNull;
 
-public class GameStopEvent extends GameEvent {
+/**
+ * Get's called, when a player has already entered an arena
+ */
+@Getter
+public class PlayerArenaJoinEvent extends PlayerEvent implements Cancellable {
 
     public final static HandlerList handlers = new HandlerList();
+    private final Game game;
+    private boolean cancelled;
 
-    public GameStopEvent(Game game) {
-        super(game);
+    public PlayerArenaJoinEvent(Player who, Game game) {
+        super(who);
+        this.game = game;
     }
 
     public static HandlerList getHandlerList() {
@@ -34,7 +46,18 @@ public class GameStopEvent extends GameEvent {
     }
 
     @Override
+    @NotNull
     public HandlerList getHandlers() {
         return handlers;
+    }
+
+    @Override
+    public boolean isCancelled() {
+        return cancelled;
+    }
+
+    @Override
+    public void setCancelled(boolean b) {
+        cancelled = b;
     }
 }
