+++ missilewars-plugin/src/main/java/de/butzlabben/missilewars/events/PlayerArenaLeaveEvent.java
@@ -16,16 +16,36 @@
  * along with MissileWars.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package de.butzlabben.missilewars.event;
+package de.butzlabben.missilewars.events;
 
 import de.butzlabben.missilewars.game.Game;
-import org.bukkit.event.Event;
+import lombok.Getter;
+import org.bukkit.entity.Player;
+import org.bukkit.event.HandlerList;
+import org.bukkit.event.player.PlayerEvent;
+import org.jetbrains.annotations.NotNull;
 
-public abstract class GameEvent extends Event {
+/**
+ * Get's called, when a player has already left an arena
+ */
+@Getter
+public class PlayerArenaLeaveEvent extends PlayerEvent {
 
+    public final static HandlerList handlers = new HandlerList();
     private final Game game;
 
-    public GameEvent(Game game) {
+    public PlayerArenaLeaveEvent(Player who, Game game) {
+        super(who);
         this.game = game;
+    }
+
+    public static HandlerList getHandlerList() {
+        return handlers;
+    }
+
+    @Override
+    @NotNull
+    public HandlerList getHandlers() {
+        return handlers;
     }
 }
