+++ missilewars-plugin/src/main/java/de/butzlabben/missilewars/game/Game.java
@@ -24,8 +24,8 @@ import de.butzlabben.missilewars.configuration.Config;
 import de.butzlabben.missilewars.configuration.PluginMessages;
 import de.butzlabben.missilewars.configuration.arena.ArenaConfig;
 import de.butzlabben.missilewars.configuration.game.GameConfig;
-import de.butzlabben.missilewars.event.GameStartEvent;
-import de.butzlabben.missilewars.event.GameStopEvent;
+import de.butzlabben.missilewars.events.GameStartEvent;
+import de.butzlabben.missilewars.events.GameStopEvent;
 import de.butzlabben.missilewars.game.enums.GameResult;
 import de.butzlabben.missilewars.game.enums.GameState;
 import de.butzlabben.missilewars.game.enums.MapChooseProcedure;
