+++ missilewars-plugin/src/main/java/de/butzlabben/missilewars/listener/PlayerListener.java
@@ -21,8 +21,8 @@ package de.butzlabben.missilewars.listener;
 import de.butzlabben.missilewars.Logger;
 import de.butzlabben.missilewars.MissileWars;
 import de.butzlabben.missilewars.configuration.Config;
-import de.butzlabben.missilewars.event.PlayerArenaJoinEvent;
-import de.butzlabben.missilewars.event.PlayerArenaLeaveEvent;
+import de.butzlabben.missilewars.events.PlayerArenaJoinEvent;
+import de.butzlabben.missilewars.events.PlayerArenaLeaveEvent;
 import de.butzlabben.missilewars.game.Game;
 import de.butzlabben.missilewars.game.GameManager;
 import de.butzlabben.missilewars.displays.MotdManager;
